requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Faker==30.8.1
pytest==8.3.3
pytest-cov==5.0.0
requests==2.28.2